metrics: always export rbd mirror state metrics
RBD mirror state might not always be primary. So we can not
use that as a check to export metrics. There are cases like
split brain or demoted image in which case we still need to
export state metrics. So, this commit removes the check.

However, other information are only available when the image
is primary and so the checks are required.

Signed-off-by: Umanga Chapagain <chapagainumanga@gmail.com>