Update newToolsDeployment according to upstream rook toolbox spec
The toolbox doesn't need to run privileged. Rook's upstream toolbox
spec hasn't had the privileged flag since Rook v1.1, which was the
basis for OCS 4.3. The privilged acces gives a warning message seen
in the OCS operator log file. So updating the newToolsDeployment
according to the upstream rook toolbox spec.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>