Fix 'CephOSDVersionMismatch' and 'CephMonVersionMismatch' alerts
Alerts, 'CephOSDVersionMismatch' and 'CephMonVersionMismatch' had a
minor issue in the query. Earlier we were counting a combination of
'namespace' and 'ceph_version', but in actual we should count
only by 'namespace' (for the second count) thus making sure we don't
have more than one ceph versions in a single namespace.

Signed-off-by: Arun Kumar Mohan <amohan@redhat.com>