odf-to-odf: fix sc validation for rbd
there was wrong check when validating sc for rbd
where ther it was checking for encryption type.
This commit change the condition from
```
(sccType == "blockpool" && !scIsFSType && sccEncryptionMethod != scEncryptionMethod)
```
to
```
(sccType == "blockpool" && !scIsFSType && sccEncryptionMethod == scEncryptionMethod)
```

this was confusion because we have `!` in fron of the first `if` which
we missed. So, basically ```(sccType == "blockpool" && !scIsFSType &&
sccEncryptionMethod == scEncryptionMethod)``` this is valid condition
for rbd and anything not matching above check is wrong sc.

Signed-off-by: subhamkrai <srai@redhat.com>