Enable rbd mirror daemon restart config automatically
Currently, the user has to set this value ceph config
set global rbd_mirror_die_after_seconds 3600 via the toolbox,
and that is not a good way for the customer as they don't have
toolbox access. We have to set this value without any user
action requirement. So setting this value in the rook-config-override
configmap.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>