Stop setting storagecluster to error phase just after creation
While storageclasses are being created it's waiting till underlying
resources like cephBlockPool, cephFilesystem & NFS are created.
In this phase, we are returning an error which results in the phase of
the storagecluster to be in Error, creating confusion for the user.

So henceforth this particular error would be caught, A requeue
Reconcile result would be given and the storagecluster would be marked
as in the Progressing Phase.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>