Define registries for Rook and Noobaa images
Lack of registry definition in this case causes the `gen-latest-csv`
command to fail when dumping them, due to the lack of an interactive TTY.

Folks may not face this issue if they already have the aliases defined in
`.cache/containers/short-name-aliases.conf`. However, it's better to be
explicit, so that newer systems that don't have such aliases don't crash
on this command.

Also refer: https://github.com/containers/image/blob/main/docs/containers-registries.conf.5.md#note-risk-of-using-unqualified-image-names

Signed-off-by: Pranshu Srivastava <rexagod@gmail.com>