Enable verbose logging when rbd-mirroring is enabled for debugging
When mirroring is enabled it creates a enable job which runs the
required ceph config set commands to enable verbose logging.
When mirroring is disabled it creates a disable job which runs the
required ceph config rm commands to disable verbose logging and deletes
both the enable and disable jobs.
RBD team needs by default verbose logging when rbd-mirroring is enabled,
to debug issues properly. This would be reverted once RDR goes to GA.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>