remove namespace deps from the bundle
- remove service account namespace ref from the role bindings

Having service Account namespace in the role bindings prevent us
from installing the operator in another namespace. Removing it will
allow us to install the operator in any namespace.

- move ocs-metrics-exporter ClusterRole and Binding to the csv

The service account namespace ref is a required field in the
ClusterRoleBinding. It can be removed from the RoleBinding but
not from the ClusterRoleBinding. Keeping the namespace ref
prevents us from installing the operator in another namespace
because the service account namespace is hardcoded.

Moving the ClusterRole to the csv allows us to remove the
namespace ref and install the operator anywhere.

- remove hardcoded namespace from the metrics-exporter deployment

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>