Warn in logs if both in-transit encryption & compression are enabled
This adds a warning in the logs if both in-transit encryption and
compression are enabled. This will help the user to understand the
behaviour.

Combining encryption with compression reduces the level of
security of messages between peers. In case both encryption and
compression are enabled, compression setting will be ignored and
message will not be compressed.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>