Refactor & cleanup the flow of rook-ceph-operator-config configmap
This configmap is purely reserved for any user overrides to be applied.
It is created by the ocsinitialization controller & we dont
reconcile it afterward as it can reset any values the customer has set.
The configmap is watched by rook-ceph operator and values set here
have higher precedence than the default values set
in the rook operator pod env vars.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>