Refactor & cleanup the flow for ocs-operator-config configmap
This configmap is reserved just for ocs operator use, primarily for
passing values to rook-ceph-operator & is not meant to be modified by
the user. Ocsinitialization controller will create this configmap
& the values are set by the storagecluster controller at the start
of the Reconcile. The needed keys from the configmap are passed to
rook-ceph operator pod as env variables. When any value in the
configmap is updated, the rook-ceph-operator pod is restarted to
pick up the new values.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>