remove the negative external cephcluster conditions
we are changing the conditions when cephcluster is in connecting state
or error state and reflecting the ceph cluster state in the
storagecluster conditions accordingly. But once ceph cluster is in
connected state we are not changing the conditions at all. Which makes
storagecluster conditions out of sync with the cephcluster when it is in
connected state. Removing the conditions will not show wrong status and
keep both the CR's in sync.

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>