Construct the topology map freshly instead of just adding to the map
We never removed the labels for deleted nodes, instead, we just added
any new information to the existing topology map. As a result, even
after a node has been deleted, its labels will still appear on the
topology map. So, it needs to be fixed by generating a fresh topology
map for each reconcile so that it will always be in sync with the
cluster's current state.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>