Refactor GetNetworkSpec to handle different network connection specs
RequireMsgr2 will be set by default if it's not an external cluster
or a provider cluster, so v2 port will be the default port.
For a new cluster, The mons will have only the v2 port.
For an upgraded cluster mons will have both v1 and v2 ports. But csi
will use the v2 port for any new volume mounts.

For external cluster or provider cluster, v1 port will be used by
default. In this case for both new cluster and for an upgraded cluster
mons will have both v1 and v2 ports. But csi will use the v1 port. If
the user wants to use Encryption, compression or v2 port, they can
specify it in the network spec. In that case csi will switch to v2 port.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>