Return ms_mode=legacy instead of blank value for External/Provider mode
In the case when for external mode upgrade happens from 4.12 to 4.13,
getCephFSKernelMountOptions returns a blank value for external mode.
It compares this with the current value of the key. But the configmap in
this case also gives a blank value for the non-existent key.
So it thinks the configmap is already in the correct state, skips
adding the key to the cm. Which results in rook-ceph-operator pod being
in CreateContainerConfigError state with the error message. So instead
of the blank value, we return the ms_mode=legacy value which has
the same effect as the blank value.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>