fix clusterclaim reconcile failure
Manager client is initialized and created once when the operator starts.
If CRDs are added to the cluster after manager client is initialized, it
can't be used for CRUD operations on resources belonging to these newly
added CRDs. To workaround that issue, we create a ClusterClaim client
dynamically when we confirm the existence of ClusterClaim CRD.

Signed-off-by: Umanga Chapagain <chapagainumanga@gmail.com>