BZ2139037 : Use strings ports in service in routes
The BZ2139037 planned to expose differet routes for RGW service for TLS
and non TLS ports. But current code uses the integer value for the
ports. Since port in mentioned in RGW service for non-secure is
different than one rgw internally uses in the pod. Hence the non-secure
RGW route was not accessible. Also changed the termination policy to
edge for the non-secure RGW route.

Signed-off-by: Jiffin Tony Thottan <thottanjiffin@gmail.com>