Use `IfNotPresent` pull policy for images pulled by digest
Currently some of the images use the `Always` image pull policy. That
isn't really necessary when the images are pulled by digest instead of
by tag, and it introduces an additional network round trip to the image
registry server that takes time and may fail. To avoid that this patch
changes the `csv-merger` tool so that in addition to merging the CSVs it
also changes the `imagePullPolicy` to `IfNotPresent` for images that are
pulled by digest.

By default this doesn't produce any change in the merged CSV, because
all the images are pulled by tag, but running the following:

```
$ export OCS_IMAGE=quay.io/ocs-dev/ocs-operator@sha256:b22c334a530dc194e758b0e9974376fe75ac1ea5044d079252aeb17081be8e0e
$ make gen-latest-csv
```

Will generate the following diff:

```diff