Always call the ensureOCSOperatorConfig func irrespective of the mode
Earlier the ensureOCSOperatorConfig func was being called while creation
of the cephcluster, but in Noobaa standalone mode it was not being
called as a cephcluster is not created in that mode. So we have to
ensure that the ensureOCSOperatorConfig func is called irrespective of
the modes. So we are moving it to the reconcile.go file and putting it
at the end of reconcilePhases func.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>