Always update the template parameters & objects to keep it up-to-date
Previously once the template existed we were not reconciling it. This
resulted in the template using very old rook-ceph images and not having
the newly added parameters. We want the template to be updated to use
new parameters, any new ceph commands, latest rook-ceph image, etc.
Trying to update the whole template object also doesn't work. So we
need to update the parameters & objects separately to keep the template
up-to-date.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>