Update the root Dockerfile & use it for building the ocs operator image
The whole repo is copied to a builder & the binaries are built there.
It makes the building process independent of the host system.
After building the binaries are copied to a lighter base image.
The makefile is updated to by default use this Dockerfile for building
the operator image. Other unnecessary files are removed.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>