Reset KMSServer connection error field when connection restablishes
KMSServerConnectionAlert is not revoked even when the KMS connection is
re-established. The alert is dependent on StorageCluster object's
'Status.KMSServerConnection.KMSServerConnectionError' field which is
populated when KMS is unreachable. But we fail to reset the field when
connection is re-established.

Fix is to reset this field when connection is fine.

Signed-off-by: Arun Kumar Mohan <amohan@redhat.com>