provider-server: enhance logging of corner case
Improve Logging for ocs-consumer Addon Re-installation with Same
Onboarding Ticket.

Previously, re-installing the ocs-consumer addon on a cluster using
the same onboarding ticket led to installation failures and left
the StorageCluster in an 'Error' phase with generic error
"Failed to get storageConsumer. storageconsumer not found".

The enhancement introduced in this commit addresses this issue by
incorporating more informative and accurate logging.

Bug: https://bugzilla.redhat.com/show_bug.cgi?id=2088434

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>