Add annotation for multus in toolbox after checking the multus network name for namespace
It is possible that the NetworkAttachmentDefinition for multus network
is configured in a different namespace than the one in which the operator is deployed
and multus is configured. In that case, the selector (eg: public) value in the
storagecluster CR is of the following pattern `NAD_namespace/NAD_name`.

Therefore, check for namespace value in the selector and add annotation to the
ceph toolbox accordingly.

BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1982721

Signed-off-by: Nikhil-Ladha <nikhilladha1999@gmail.com>