Add deployment check for rgw gateway pods
This commit adds a check for the number of ready deployment of
rook-ceph-rgw-* pods, this change is done because, ODF/Rook used to run a routine that regularly created a bucket and then wrote/read the bucket to test the RGW health, now the status checking is removed. We now need to reflect the "Readyness" and the "Connected" nature of status of the CephObjectStore.

Signed-off-by: Divyansh Kamboj <dkamboj@redhat.com>