rook: generate rook csv using local script
Let's generate rook csv using local bash script that will
be inside ocs-operator code and fetch from rook image. It gives us
advantage of not depending on rook image for their csv and we can pass
branch of commit for specific csv. Also for rook, it doesn't make sense
to have csv codes/fiels in rook upstream and the rook csv is used
only for the downstream. So, in future rook will remove the csv files
from rook image.

Signed-off-by: subhamkrai <srai@redhat.com>