Fix Provisioner Namespace Configuration
Previously the storage cluster was deployed within the operator
namespace, which was matching the Provisioner namespace and working
well. However with the introduction of multiple StorageClusters, using
the storagecluster namespace as provisioner namespace is no longer
viable. Instead we should use operator namespace as that is the
namespace where provisoners are deployed.

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>