Use context.TODO() instead of reconciler's context
The reconciler's context can be `nil`, when the handler is executed
if the storagecluster CR is not present, as the reconcile loop will
not be triggered and the context field will never be initialised.

But, even if it's `nil` the `r.Client.List` operation doesn't always
panic because there is a race scenario of the `cache` being synced by
the time code execution reaches here. If the cache is not synced, then
it waits for the cache to sync before executing the `List` operation
and sends a close request via the `Done` channel of `ctx`, which results
in a panic as `ctx` is nil in this case.

Therefore, we should not be using the `r.ctx` field and instead use the
`context.TODO()` value to execute the `List` operation.

Signed-off-by: Nikhil-Ladha <nikhilladha1999@gmail.com>