Separate ceph toolbox from ocsinitialization
With the development for multiple storagecluster happening in 4.15, we
need to separate ocsinitialization from the cephtoolboox. This change
is needed because there will be only one ocsinitialization CR in the
operator's namespace, and no ocsinitialization CR might exist in the
namespaces where storage clusters are deployed. Also in case we have
multiple storageclusters we might not want the toolbox to be enabled
for all of them.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>