Fix test TestIsActiveStorageCluster
The previous version of the list function was only listing storage
clusters within the same namespace as the instance. Consequently,
it was not listing the empty storage cluster created in test case 4,
and the test case was passing. However, with the current implementation
that lists all namespaces, we need to ensure that we skip creating the
empty storage cluster.

Signed-off-by: Nitin Goyal <nigoyal@redhat.com>