skip checking osd store type for external clusters
Currently there is no way to determine if external clusters are
optimized for DR, that is, they are bluestore-rdr. So we should
skip checking for the OSD store type in the cephCluster status in
case of a external cluster.

Note: We would want to revisit this in future once we start
supporting external clusters for RDR.

Signed-off-by: sp98 <sapillai@redhat.com>