fix lock when PV is resynced
PV resync obtains the lock to PV cache but then delegates the cache
update job to another function which also tries to gain lock but fails.
This commit separates the locking from cache update function to a shared
function without lock. This new function is not thread safe and should
be called from within threadsafe function.

Signed-off-by: Umanga Chapagain <chapagainumanga@gmail.com>