refactor getNodeNameForPV in metrics-exporter
This commit refactors getNodeNameForPV function to retrieve the node on
which the PV is created by inspecting the Node objects instead of the
Pod object.

Inspecting Nodes is faster than inspecting Pods when number of volumes
in the cluster is less than the number of Pods in the PVC namespace.
Volumes in the cluster = Avg number of volume per Node * Number of Nodes

Signed-off-by: Umanga Chapagain <chapagainumanga@gmail.com>