refactor ceph client creation for metrics-exporter
This commit adds flexibility to create Ceph client for metrics-exporter
usage. If user provides a user name and key for a ceph cluster, we will
use that to create a ceph client for that cluster. If user does not
provide a separate user name and key, we will piggyback on the ceph-csi
rbd provisioner auth token which has the same access as required for
monitoring. If that is also not found, we will report an error.

To be able to create the client, we need to know the namespace of the
ceph cluster as well as the namespace where the auth tokens for the
cluster is created. These can be passed via command line args during
exporter deployment.

Signed-off-by: Umanga Chapagain <chapagainumanga@gmail.com>