add CSI_DRIVER_NAME_PREFIX to Rook config
The current practice is to prefix the names
of CephCSI drivers with the namespace name
of the deployed Rook operator. However, if
Rook is deployed in a different namespace
(e.g. "xyz"), then the CephCSI driver name
will also be prefixed with "xyz". This is not
desirable, as we always want to use
"openshift-storage" as the prefix, regardless
of the namespace where Rook is deployed. This
PR sets the CSI_DRIVER_NAME_PREFIX to a
static value, i.e. "openshift-storage"
for all CephCSI drivers.

Signed-off-by: Madhu Rajanna <madhupr007@gmail.com>