provider: avoid CSI secret name collisions
Previously, when fetching a storage claim config, a remote consumer was
given information for creating the Ceph CSI Secrets that was identical
to the secret names on the provider side. Running a consumer in the same
namespace as the provider introduced namespace collisions.

Hilariously, by virtue of being the same information, things just kinda
worked out anyway. Still, this is likely going to produce massive
problems later, so better to fix it now.

Signed-off-by: Jose A. Rivera <jarrpa@redhat.com>