provider: clarify variable names in storageclassrequest manager
Probably as a holdover from a previous design, the provider server code
uses "storageClassRequestName" in multiple places where it is not
appropriate. These often hold the name for the *StorageClassClaim* that
came in as part of the FulfillStorageClassClaim() call. I seriously
lost hours of my time before I figured out this mistake, so I'm saving
everyone else the headache in the future.

Signed-off-by: Jose A. Rivera <jarrpa@redhat.com>