makefile: use more specific paths for controller-gen
Previously, the paths given to controller-gen have it scanning all go
packages in the repo for relevant kubebuilder markers. This can cause
problems with parts of the repo that may need additional libraries
installed in the runtime environment but are not required for the
operator or its bundle manifests.

Limiting controller-gen to only scan the directories that are actually
relevant to its functionality resolves this, and just seems like a good
idea in general.

Signed-off-by: Jose A. Rivera <jarrpa@redhat.com>