do not check for storageconsumer when storagerequest is being deleted
there is a possibility that getting storageconsumer from cache would
work even when the resource is deleted however if the event gets
requeued then we will always hit failure while trying to get deleted
resource after a cache sync.

with this commit we only check presence of storageconsumer if
storagerequest is not being deleted as the deletion of storagerequest is
generally triggered via owner references and storageconsumer is owner
of storagerequest.

Signed-off-by: Leela Venkaiah G <lgangava@ibm.com>