Set the deviceClasses of the pools to avoid wrong data placement
Earlier when replica-1 was getting enabled on an existing cluster the
pool was setting it's deviceClass to replicated but there were no osds
with this deviceClass as osd's during normal creation were given ssd
deviceClass if blank, and now they can't change. To avoid & solve this
problem we have to intelligently set the deviceClass for the pools,
Depending on whether replica-1 is enabled & the number of deviceClasses
found in the cephCluster CR status.

To do this we added a DefaultCephDeviceClass field in StorageCluster
Status which will  hold the default ceph device class to be used for
the pools. Depending  on whether replica-1 is enabled & the number of
deviceClasses found in  the status this value is determined and set in
the status.

This also includes the unit test as well.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>