Fix pg_autoscaler won't scale due to overlapping roots
There was no CephBlockPool defined for the .mgr pool, therefore it was
using the default crush rule (with no deviceClass specified).
Also the CephObjectStore & CephFilesystem metadataPool does not define
the deviceClass (it is only set in the dataPools).

To fix this Create a CephBlockPool CR for the .mgr pool with the
same poolSpec as the default CephBlockPool. Also set the deviceClass
for the metadataPool in CephObjectStore and CephFilesystem.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>