Always enforce the TSCs even if the placement is specified by the user
Currently TSCs are enforced only if there is no other placement spec
defined for the osds & osd-prepares. Without the TSCs, the OSDs will
rarely be balanced. So we should always enforce the TSCs even if the
placement is specified by the user.

This also refactors the code relating to osd placement to make it more
readable and maintainable.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>