Expose storageCluster flag to allow updating of the OSD device classes
Earlier the device class of an OSD was determined at provisioning and
it was not possible to update thereafter. In some scenarios the admin
may want to force update the device class to a new value.
To avoid accidental updates, rook requires first setting
allowDeviceClassUpdate in the storage spec of the cephcluster for which
we are exposing a new flag at
Spec.ManagedResources.CephCluster.AllowDeviceClassUpdate.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>