Use msgrv2 always for all ODF modes(internal,external,provider)
Initially in 4.13 when encryption in-transit was implemented we started
to use msgrv2 for internal mode only. External mode & Provider mode were
excluded.

For external mode the concern was the availability of msgrv2
port always on the RHCS cluster. But now we have reached ceph v18 &
msgrv2 is being by default on since ceph Nautilus 14.2. So we
can now use mgsr2 for external mode.

For provider mode the concern was availability of the required kernel
versions on old client clusters. But now that concern also seems no
longer valid. So we can now use msgrv2 for provider mode as well.

Also as we would always use msgrv2, we can now simplify the
GetCephFSKernelMountOptions function to remove the conditional checks.
If encryption is enabled use "secure" otherwise "prefer-crc".

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>