Remove check of object type & return in enqueueStorageClusterRequest
The enqueueStorageClusterRequest event handler is used to enqueue a
reconcile request for the controller from the watched objects.
Although the handle is being used by all the watched objects, but
it checks the object type and returns if the object is not of type
OCSInitialization. Due to this when any thing happens with the watched
objects other than OCSInitialization, the reconcile request is not
enqueued and the controller does not reconcile the object.

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>